:bug: :: 숫자 버그 수정

+++ src/main/kotlin/com/eathub/eathub/domain/food/application/presentation/dto/FoodStatsMessage.kt
@@ -2,6 +2,6 @@ package com.eathub.eathub.domain.food.application.presentation.dto
 
 class FoodStatsMessage(
     val usedAmount: Long,
-    val amountPerPerson: Int,
+    val amountPerPerson: Long,
     val remainedAmount: Long
 )
